parser-produit/requirements.txt
fastapi
uvicorn[standard]
pydantic
psycopg2-binary
SQLAlchemy
pytesseract
Pillow
beautifulsoup4
python-multipart
